tests/perf: verify new options for 64bits gpu timestamps
The timestamp registers are 64bits wide. The reports from the OA unit
only include the least significant 32bits. We added a new option in
the perf driver to add back a new 64bits field in the i915 perf
reports containing the full timestamp.

We can verify that the timestamps match on their lower 32bits parts.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>