benchmark: add gem_exec_delay
This benchmark's goal is to measure the time it takes between a
request submitted to the hardware (either from i915 or guc) and when
the first RCS command stream instruction from the user batchbuffer is
executed.

This benchmark works by monitoring the RCS using the OA unit. The OA
unit on Gen8+ generates a report every time an execlists port is
written to (or legacy ringbuffer). We then measure the delta between
the RCS timestamp that is captured using MI_STORE_REGISTER_MEM (which
is the first and only instruction in our batchbuffer) and the previous
execlists port write.

The benchmark alternates between 2 contexts and builds dependencies
between their execbufs to make sure execlists ports writes are never
combined (due to similar priority when the scheduler picks up requests
to submit).

At the moment the benchmark provides 2 mode of operations :

   - Alternate between 2 contexts of similar default configurations
   - Alternate between 2 contexts of different powergating
     configurations (default & 1slice/1subslice)

On Skylake GT4 (3 slices) :

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m none
          average=22.73us min/max=20.17us/36.17us variance=2.11ms std_deviation=1.45us

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m rpcs
          average=72.93us min/max=60.75us/87.42us variance=84.41ms std_deviation=9.19us

On Kabylake GT3 (2 slices) :

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m none
          average=22.64us min/max=21.00us/24.42us variance=114.86us std_deviation=338.91ns

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m rpcs
          average=67.90us min/max=59.08us/89.42us variance=62.42ms std_deviation=7.90us

On Broxton 2x6 (1 slice) :

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m none
          average=38.88us min/max=35.94us/68.23us variance=23.93ms std_deviation=4.89us

     $ sudo ./build/benchmarks/gem_exec_delay_bench -r 100 -m rpcs
          average=37.57us min/max=34.17us/63.59us variance=35.06ms std_deviation=5.92us

On Cannonlake GT0.5 (1 slice) :

     $ sudo ./build/benchmarks/gem_exec_delay_bench -m none -r 100
          average=36.31us min/max=35.67us/41.75us variance=683.09us std_deviation=826.49ns

     $ sudo ./build/benchmarks/gem_exec_delay_bench -m rpcs -r 100
          average=36.11us min/max=35.17us/37.17us variance=106.81us std_deviation=326.82ns

Subslice powergating doesn't seem to have an measurable impact, I do
wonder though whether it's actually working.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>