tests/perf: add a command streamer coherency test
We would like to verify that we can copy a report generated by the OA
unit using the command streamer. It's not obvious how the different
units communicate and what are the restrictions.

So far it seems we don't have any restriction, but in case that
changes we would like to be warned.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>