lib: fix aubdump dependency in autotools build
In commit 1bb318b32db003a ("lib: create new dependency for aubdump")
we created a new dependency for the aubdump tool, so it doesn't pull
the whole of libigt into some other process' address space. This
previous change only dealt with meson and I forgot to modify the
autotools build.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>